Add musicstore-catalog service

+++ M7-Challenge-Netzel-John/musicstore-catalog/src/main/java/com/company/musicstorecatalog/controller/MusicStoreAlbumController.java
@@ -3,9 +3,11 @@ package com.company.musicstorecatalog.controller;
 import com.company.musicstorecatalog.model.Album;
 import com.company.musicstorecatalog.repository.AlbumRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 
 @RestController
 @RequestMapping(value = "/album")
@@ -24,17 +26,24 @@ public class MusicStoreAlbumController {
     }
 
     @PostMapping
+    @ResponseStatus(HttpStatus.CREATED)
     public Album createLabel(@RequestBody Album label) {
         return albumRepo.save(label);
     }
 
     @DeleteMapping(value = "/{id}")
+    @ResponseStatus(HttpStatus.NO_CONTENT)
     public void deleteLabel(@PathVariable Integer id) {
+        if(!albumRepo.findById(id).isPresent()) // Check to make sure record with ID exists
+            throw new NoSuchElementException("Attempting to delete record that does not exist:" + id);
         albumRepo.deleteById(id);
     }
 
     @PutMapping(value = "/{id}")
+    @ResponseStatus(HttpStatus.NO_CONTENT)
     public void updateLabel(@PathVariable Integer id, @RequestBody Album label) {
+        if(!albumRepo.findById(id).isPresent()) // Check to make sure record with ID exists
+            throw new NoSuchElementException("Attempting to update record that does not exist:" + id);
         albumRepo.save(label);
     }
 }
